requirements.txt
requests==2.28.1
openai==0.27.0
python-dotenv==0.21.0
